syncer: split out controller code from syncer.go
